feat: Add XValidations in the generated documentation
With Kubernetes 1.25, CEL will be added, as an alternative to enforcing
validations, rather than webhooks. This allows, among other things,
the ability to enforce immutability on fields.

Since this is documented in the CRD YAML itself, bubbling this up
into the documentation allows one to see custom validation rules

An example kubebuilder comment in Go would be:

```
// +kubebuilder:validation:XValidation:rule="self == oldSelf", message="Value is immutable"
```

This would result in the following in the CRD:

```
x-kubernetes-validations:
- message: Value is immutable
  rule: self == oldSelf
```

And now the generated documentation will reflect this:

```
<i>Validations</i>:<li>self == oldSelf: Value is immutable</li>
```

Signed-off-by: Trynity Mirell <trynity@mirell.org>